internal/repository/account_test.go
package repository

import (
	"context"
	"sbank/internal/controller/dto"
	"testing"

	"github.com/stretchr/testify/require"
)

func TestCreateAccount(t *testing.T) {
	arg := dto.CreateAccountDTO{
		Owner:    "tom",
		Balance:  100,
		Currency: "USD",
	}

	account, err := testRepository.CreateAccount(context.Background(), arg)

	require.NoError(t, err)
	require.NotEmpty(t, account)

	require.Equal(t, arg.Owner, account.Owner)
	require.Equal(t, arg.Balance, account.Balance)
	require.Equal(t, arg.Currency, account.Currency)

	require.NotZero(t, account.ID)
	require.NotZero(t, account.CreatedAt)
}
